Add fixes for rpm and add a flag to disable log collection

+++ package-tooling/linux/rpm/mw-agent.spec
@@ -6,7 +6,7 @@ License: GPL
 Group: Development/Tools
 Source0: %{package_name}-%{release_version}-%{arch}.tar.gz
 Provides: %{package_name}
-Obsoletes: %{package_name} <= %{release_version}
+Obsoletes: %{package_name} < %{release_version}
 
 %description
 Middleware Agent(%{package_name}) service enables you to monitor your infrastructure and applications.
@@ -40,13 +40,21 @@ chmod u+x /opt/%{package_name}/.postinstall.sh
 /opt/%{package_name}/.postinstall.sh
 
 %preun
-systemctl stop %{package_name}
-systemctl disable %{package_name}
+if [ $1 -gt 0 ]; then
+    echo "Upgrade in progress, skipping pre-uninstallation steps."
+else
+    systemctl stop %{package_name}
+    systemctl disable %{package_name}
+fi
 
 %postun
-rm -f /etc/%{package_name}/agent-config.yaml
-rm -f /etc/%{package_name}/otel-config.yaml
-rmdir  /etc/%{package_name}
-rmdir /opt/%{package_name}/bin
-rmdir /opt/%{package_name}
+if [ $1 -gt 0 ]; then
+    echo "Upgrade in progress, skipping post-uninstallation steps."
+else
+    rm -f /etc/%{package_name}/agent-config.yaml
+    rm -f /etc/%{package_name}/otel-config.yaml
+    rmdir  /etc/%{package_name}
+    rmdir /opt/%{package_name}/bin
+    rmdir /opt/%{package_name}
+fi
